requirements.txt
diff-match-patch==20181111
Flask==1.1.2
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-RQ2==18.3
Flask-SQLAlchemy==2.4.3
https://github.com/hasgeek/baseframe/archive/master.zip
https://github.com/hasgeek/coaster/archive/master.zip
https://github.com/hasgeek/flask-lastuser/archive/master.zip
premailer==3.7.0
pystache==0.5.4
short_url==1.2.2
